+++ app/models/export/tasks/row.rb
@@ -11,13 +11,17 @@ module Export
         "#{task.period_year}-#{format('%02d', task.period_month)}"
       end
 
+      def status
+        task.status == 'correcting' ? 'completed' : task.status
+      end
+
       def row_values
         [
           task.id,
           year_and_month,
           task.supplier.salesforce_id,
           task.framework.short_name,
-          task.status,
+          status,
           FIXED_TASK_TYPE,
           EMPTY_FOR_NOW,
           EMPTY_FOR_NOW