+++ app/models/task.rb
@@ -5,9 +5,17 @@ class Task < ApplicationRecord
     state :unstarted, initial: true
     state :in_progress
     state :completed
+    state :correcting
 
     event :completed do
-      transitions from: %i[unstarted in_progress completed], to: :completed
+      transitions from: %i[unstarted in_progress correcting completed], to: :completed
+    end
+
+    event :cancel_correction do
+      before do
+        destroy_incomplete_correction_submissions if correcting?
+      end
+      transitions from: :correcting, to: :completed
     end
   end
 
@@ -26,6 +34,7 @@ class Task < ApplicationRecord
     order(Arel.sql("CASE submissions.aasm_state WHEN 'completed' THEN 1 ELSE 2 END"), created_at: :desc)
   end
   has_one :active_submission, completed_or_latest_scope, class_name: 'Submission', inverse_of: :task
+  has_one :latest_submission, -> { order(created_at: :desc) }, class_name: 'Submission', inverse_of: :task
 
   def file_no_business!(user)
     transaction do
@@ -47,6 +56,16 @@ class Task < ApplicationRecord
 
   # Returns true when the task is yet to be completed by the Supplier
   def incomplete?
-    !completed?
+    !completed? && !correcting?
+  end
+
+  def destroy_incomplete_correction_submissions
+    transaction do
+      submissions.where('created_at > ?', active_submission.created_at).find_each do |submission|
+        submission.entries.destroy_all
+        submission.files.destroy_all
+        submission.destroy
+      end
+    end
   end
 end